cleanup codebase and update manual
