+++ README.md
@@ -27,25 +27,6 @@ You can refer to `kup -h` for more options. For example,
 ### Using `KMIR`
 
 At the current stage, `KMIR` is to be used as a command-line tool. We list the main functions in the following, while users are recommended to explore the latest options using `kmir -h` or more options of a command using `kmir {cmd} -h`.
-
-  ```bash
-  usage: kmir [-h] {parse,run,prove,show-kcfg,view-kcfg} ...
-
-  KMIR command line tool
-
-  positional arguments:
-    {parse,run,prove,show-kcfg,view-kcfg}
-                          Command to execute
-      parse               Parse a MIR file
-      run                 Run a MIR program
-      prove               Prove a MIR specification WARN: EXPERIMENTAL AND WORK IN PROGRESS
-      show-kcfg           Display tree show of CFG
-      view-kcfg           Display tree view of CFG
-
-  options:
-    -h, --help            show this help message and exit
-  ```
-
 #### Parsing a MIR file
 The `kmir parse` command invokes the parser generated by K and outputs a KAST representation of MIR abstract syntax parsed from a MIR file. 
 
@@ -53,8 +34,6 @@ For example,
   ```sh
   kmir parse demo/assert-true.mir
   ```
-
-Users can explore more options by running `kmir parse -h`.
 
 #### Execute a MIR program
 The `kmir run` command will interpret a MIR program according to the predefined operational rules of `MIR`. This command invokes the LLVM interpreter generated by K and execute the MIR code in the file. 
@@ -67,20 +46,20 @@ kmir run demo/assert-true.mir
 #### Prove a MIR program 
 **(Under development, command interface unstable )**
 
-Use the following commands from the `kmir` directory to manually prove `k` specification files for MIR.
+KMIR provides the command `kmir prove` to prove the claims in the provided specification file.
 ```sh
 kmir prove demo/simple-spec.k
 ```
-This command will automatically store the proof graph- `kcfgs` in the same directory as the `spec-file`. User could specify a specific directory using the flag `--save-directory`.
+This proof results are by default stored in the same directory as the `spec-file`. User could specify a specific directory using the flag `--save-directory`.
 
-#### Display the tree show of a specified proof
+#### Display a proof in the tree view
 ```sh
-kmir show-kcfg SIMPLE-SPEC.statement.1 --proof-dir demo
+kmir show-proof SIMPLE-SPEC.statement.1 --proof-dir demo
 ```
 
-#### Display an interative view of a specified proof
+#### Display a proof in an interative tree format
 ```sh
-kmir view-kcfg SIMPLE-SPEC.statement.1 --proof-dir demo
+kmir view-proof SIMPLE-SPEC.statement.1 --proof-dir demo
 ```
 
 ## Build from source
