+++ kmir/src/kmir/__init__.py
@@ -1,5 +1,8 @@
 from typing import Final
 
-# from .kmir import KMIR
+from .kmir import KMIR
+from .parse import parse
+from .prove import prove, show_proof, view_proof
+from .run import run
 
-VERSION: Final = '0.2.19'
+VERSION: Final = '0.2.21'
