+++ kmir/src/kmir/utils.py
@@ -1,17 +1,11 @@
 import logging
-import sys
-import traceback
-from collections.abc import Callable
+import re
 from pathlib import Path
-from typing import Final, TypeVar
+from typing import Callable, Final, TypeVar
 
 from pyk.cterm import CTerm
-from pyk.kast.inner import KInner, Subst
 from pyk.kast.manip import set_cell
 from pyk.kast.outer import KApply, KClaim, KRewrite, KSequence
-from pyk.kcfg import KCFG, KCFGExplore
-
-from .preprocessor import preprocess
 
 T1 = TypeVar('T1')
 T2 = TypeVar('T2')
@@ -20,8 +14,35 @@ NodeIdLike = int | str
 
 _LOGGER: Final = logging.getLogger(__name__)
 
+ALLOC_REFERENCE = r'#\(-*alloc[0-9]+(?:\+0x[0-9a-fA-F]+)?-*\)#'
+BYTE_VALUE = '[0-9a-fA-F][0-9a-fA-F]'
+UNINITIALIZED_BYTE = '__'
+ALLOC_ITEM = '|'.join([ALLOC_REFERENCE, BYTE_VALUE, UNINITIALIZED_BYTE])
+ALLOC_VALUE = r'(\s*(?: (' + ALLOC_ITEM + '))+)'
+ALLOC_SUFFIX = r'\s+│.*$'
+
+HEX_CLEANUP_SUFFIX = re.compile(r'^' + ALLOC_VALUE + ALLOC_SUFFIX)
+HEX_CLEANUP_SEPARATOR = re.compile(r'^(\s+0x[0-9a-fA-F]+\s+)│' + ALLOC_VALUE + ALLOC_SUFFIX)
+
 
 def preprocess_mir_file(program_file: Path, temp_file: Path) -> None:
+    def cleanup_hex_dump(line: str) -> str:
+        line = line.replace('╾', '#(').replace('─', '-').replace('╼', ')#')
+        m = HEX_CLEANUP_SUFFIX.match(line)
+        if not m:
+            m = HEX_CLEANUP_SEPARATOR.match(line)
+            if not m:
+                return line
+            return '%s|%s' % (m.group(1), m.group(2))
+        return m.group(1)
+
+    def process_line(line: str) -> str:
+        line = cleanup_hex_dump(line)
+        return line
+
+    def preprocess(program_text: str) -> str:
+        return '\n'.join(process_line(line) for line in program_text.splitlines())
+
     temp_file.write_text(preprocess(program_file.read_text()))
 
 
@@ -38,26 +59,6 @@ def ensure_ksequence_on_k_cell(cterm: CTerm) -> CTerm:
         _LOGGER.info('Introducing artificial KSequence on <k> cell.')
         return CTerm.from_kast(set_cell(cterm.kast, 'K_CELL', KSequence([k_cell])))
     return cterm
-
-
-def process_exception(err: Exception, curr_output: list[str], err_msg: str) -> None:
-    curr_output.append(err_msg)
-    print(err, flush=True, file=sys.stderr)
-    print(traceback.format_exc(), flush=True, file=sys.stderr)
-
-
-def print_model(node: KCFG.Node, kcfg_explore: KCFGExplore) -> list[str]:
-    res_lines: list[str] = []
-    result_subst = kcfg_explore.cterm_get_model(node.cterm)
-    if type(result_subst) is Subst:
-        res_lines.append('  Model:')
-        for var, term in result_subst.to_dict().items():
-            term_kast = KInner.from_dict(term)
-            res_lines.append(f'    {var} = {kcfg_explore.kprint.pretty_print(term_kast)}')
-    else:
-        res_lines.append('  Failed to generate a model.')
-
-    return res_lines
 
 
 def node_id_like(s: str) -> NodeIdLike:
